fix: split release assets into chunks (2 GiB limit)
note: I was not careful reading the GitHub Docs, but there is a [2 GiB
limit](https://docs.github.com/en/repositories/releasing-projects-on-github/about-releases#storage-and-bandwidth-quotas)
for releases which I hit in [a recent
build](https://github.com/johnrichardrinehart/nix/runs/4174613064?check_suite_focus=true).
The modifications to the GitHub Action are an attempt to ameliorate
that.