fix(/laptop.nix): try to fix systemd failure at local-fs.target
refs:
[1] https://github.com/systemd/systemd/commit/3519d230c8bafe834b2dac26ace49fcfba139823
[2] https://copyninja.info/blog/systemd_automount_entry.html
I had two ways to go, here. I could try x-systemd.requires=<path> or I
could try x-systemd.requires-mounts-for=<path>. They both are documented
to add both a Require= and After= statement in the generated .mount
target.

The latter still tries to enforce this dependency even if
`noauto` is specified on the dependent path (the one we need). So, if,
for some reason, any of the dependencies of my overlay are marked as
`noauto` and aren't started, then I understand my overlay mounts may
fail. This isn't the best. I want systemd to try to apply these overlays
regardless of the status of the dependencies. It will fail if they're
not available by the nature of the overlay. There's no need to introduce
another potential point of failure.

So, even though the commit message in systemd which introduces this
change (3519d230c) gives an overlay filesystem example using
x-systemd.requires-mounts-for, I think that x-systemd.requires is
preferable. And, the manpage documentation added to the same commit
indicates that x-systemd.requires is useful in the case of overlay
mounts.
> This option is particularly useful for mount point declarations that
> need an additional device to be around (such as an external journal
> device for journal file systems) or an additional mount to be in place
> (such as an overlay file system that merges multiple mount points).