fix: remove pulseaudio
I was having a problem trying to install pulseaudio (to get pactl so I
can use volume buttons). But, the nix pulseaudio package defaults to
installing and enabling systemd socket/service units which conflict with
pipewire. Trying to override the `useSystemd = true` attribute in the
package specification resulted in a failed installation of pulseaudio
(https://github.com/NixOS/nixpkgs/issues/158046), which fails the build,
generally. So, we are going to remove this dependency and have no audio
buttons until a solution is found for this problem.