chore: stripping out the cruft
This branch is to document a problem that I encountered when trying to
apply an override to a dbeaver build dependency (maven-deps). The
attribute that I wanted to override was inside of a derivation
(maven-deps') inside of a derivation (dbeaver's). Apparently, the
recommended way to override an attribute of a derivation is to use
overrideAttrs. I tried applying that in the home manager home.packages
list, but that failed.  So, I tried applying that override in an overlay
in the home manager nixpkgs.overlay. That failed. So, I tried overriding
in environment.systemPackages (even though I don't want dbeaver
installed at the system level). That failed. So, I tried overlaying
that override in the nixpkgs.overlays of the system configuration. That
failed. Then (since I'm trying to override a specific SHA value) I tried
using different SHA values and, to my suprise, different behavior was
observed with lib.fakeSha256 than with the one recommended by the build
failure. Using lib.fakeSha256: The build failure properly reported that
I was using SHA 000...00000. I converted the "sha-<base64-encoding>" to
the corresponding hexadecimal string representation and tried that, too.
I can't remember if that built correctly. But, basically, I wanted to
generate a table of derivations which I could use to test the following
combinations:

(lib.fakeSha256, hexShaRepresentation, base64ShaRepresentation) X
(home-manager override, home-manager overlay, system override, system
overlay)

That would make 12 derivations plus I wanted a couple of derivations
where I don't try to tweak the SHA (so, no override/overlay). But, I got
stuck in trying to generate the 12 above derivations. My map/fold powers
are too weak.

Every file in this branch is more/less the way I want it except for ll.
63-112. I want that generated from the data structure dictating for which
"cases" we want to generate derivations.