feat: bump nixpkgs input
This input looks a little weird - fair. But, the reason it points to
some probably-garbage-collected revision is because it _used_ to point
ot a branch, which my /run/current-system (latest system configuration)
had been build and which has since been rebased (rock-5c-nixos-25.11). I
was building the rebased version of this using my friend Justin's
machine as a remote builder. But, he asked me to stop using it for 24h
because he has some deadline and needs to do a large upload and his
upload bandwidth is pretty limited (so, the copies of build artifacts to
and from his machine is chewing into his usage). Considering I can
probably bump/build this in 24h-48h I'm just stashing the work, here,
now.