feat: do not lock and suspend on lid-close
niri appears to enter a suspend loop regardless of lid open/close. Like,
I'll `systemctl suspend` and then click the power button to resume, or
whatever, and then (9 times out of 10) the system will immediately
suspend before I have the chance to type any/all of my password into
hyprlock. I commented this configuration out on a lark and haven't seen
the problem since. So, my guess, for now, is that niri is acting on some
lid close/open event which is sent when resuming from suspend. But,
that's kind of a weird idea, so I'm not exactly sure, really.